#8 - Changed Force -> PerturbingForce. Changed Thrust -> Force, as it should have been. Force model changed compute_force -> compute_perturbation which changed output from an accel vec to a structure with both an accel vec and a torque vec. Reorganized some stuff
