+++ astrea/astro/astro/astro.hpp
@@ -54,12 +54,12 @@
 
 #include <astro/propagation/analytic/LambertSolver.hpp>
 
-#include <astro/propagation/force_models/AtmosphericForce.hpp>
-#include <astro/propagation/force_models/Force.hpp>
+#include <astro/propagation/force_models/instances/AtmosphericForce.hpp>
+#include <astro/propagation/force_models/PerturbingForce.hpp>
 #include <astro/propagation/force_models/ForceModel.hpp>
-#include <astro/propagation/force_models/NBodyForce.hpp>
-#include <astro/propagation/force_models/OblatenessForce.hpp>
-#include <astro/propagation/force_models/SolarRadiationPressure.hpp>
+#include <astro/propagation/force_models/instances/NBodyForce.hpp>
+#include <astro/propagation/force_models/instances/OblatenessForce.hpp>
+#include <astro/propagation/force_models/instances/SolarRadiationPressure.hpp>
 
 #include <astro/propagation/equations_of_motion/CowellsMethod.hpp>
 #include <astro/propagation/equations_of_motion/EquationsOfMotion.hpp>
