+++ astrea/astro/astro/propagation/equations_of_motion/CowellsMethod.cpp
@@ -52,7 +52,7 @@ OrbitalElementPartials CowellsMethod::operator()(const State& state, const Vehic
     const auto muOverRadiusCubed = mu / (R * R * R);
 
     // Run find functions for force model
-    const AccelerationVector<frames::earth::icrf> accelPerts = forces->compute_forces(state, vehicle);
+    const auto [accelPerts, torquePerts] = forces->compute_perturbations(state, vehicle);
 
     // Get vehicle-produced accels
     const AccelerationVector<frames::earth::icrf> accelVehicle = vehicle.get_command_acceleration(state);
