+++ astrea/astro/astro/propagation/force_models/PerturbingForce.test.cpp
@@ -18,7 +18,7 @@
 
 #include <astro/frames/CartesianVector.hpp>
 #include <astro/platforms/Vehicle.hpp>
-#include <astro/propagation/force_models/Force.hpp>
+#include <astro/propagation/force_models/PerturbingForce.hpp>
 #include <astro/state/State.hpp>
 #include <astro/state/orbital_elements/instances/Cartesian.hpp>
 #include <astro/systems/AstrodynamicsSystem.hpp>
@@ -31,12 +31,12 @@ using namespace mp_units;
 using mp_units::si::unit_symbols::km;
 using mp_units::si::unit_symbols::s;
 
-class DummyForce : public Force {
+class DummyForce : public PerturbingForce {
   public:
     DummyForce() = default;
-    AccelerationVector<frames::earth::icrf> compute_force(const State& state, const Vehicle& vehicle) const override
+    Perturbation compute_perturbation(const State& state, const Vehicle& vehicle) const override
     {
-        return AccelerationVector<frames::earth::icrf>(0.0 * km / (s * s), 0.0 * km / (s * s), 0.0 * km / (s * s));
+        return { .force = AccelerationVector<frames::earth::icrf>(0.0 * km / (s * s), 0.0 * km / (s * s), 0.0 * km / (s * s)) };
     }
 };
 
@@ -63,7 +63,7 @@ TEST_F(ForceTest, DefaultConstructor) { ASSERT_NO_THROW(DummyForce()); }
 TEST_F(ForceTest, ComputeForce)
 {
     const State state(cart, date, sys);
-    auto accel = force.compute_force(state, vehicle);
+    auto [accel, torque] = force.compute_perturbation(state, vehicle);
     ASSERT_EQ(accel.get_x(), 0.0 * km / (s * s));
     ASSERT_EQ(accel.get_y(), 0.0 * km / (s * s));
     ASSERT_EQ(accel.get_z(), 0.0 * km / (s * s));
