+++ astrea/astro/tests/nasa_6dof_checkcases/orbital/driver.cpp
@@ -35,11 +35,11 @@
 #include <astro/propagation/equations_of_motion/KeplerianVop.hpp>
 #include <astro/propagation/equations_of_motion/TwoBody.hpp>
 
-#include <astro/propagation/force_models/AtmosphericForce.hpp>
+#include <astro/propagation/force_models/instances/AtmosphericForce.hpp>
 #include <astro/propagation/force_models/ForceModel.hpp>
-#include <astro/propagation/force_models/NBodyForce.hpp>
-#include <astro/propagation/force_models/OblatenessForce.hpp>
-#include <astro/propagation/force_models/SolarRadiationPressure.hpp>
+#include <astro/propagation/force_models/instances/NBodyForce.hpp>
+#include <astro/propagation/force_models/instances/OblatenessForce.hpp>
+#include <astro/propagation/force_models/instances/SolarRadiationPressure.hpp>
 #include <astro/propagation/numerical/Integrator.hpp>
 
 #include <astro/astro.macros.hpp>
