+++ astrea/trace/trace/platforms/sensors/fov/instances/PolygonalFieldOfView.hpp
@@ -91,9 +91,9 @@ class PolygonalFieldOfView : public FieldOfView {
     // Probably will be some body-fixed frame aligned with the sensor boresight
     // The key angle is the azimuthal angle around the boresight, and the value angle is the
     // off-boresight angle at that azimuth
-    gtl::btree_map<Angle, Angle> _points; //<! Map of angles defining the polygonal field of view
-    Angle _minHalfAngle = 0.0 * mp_units::angular::unit_symbols::rad; //<! Minimum off-boresight half angle
-    Angle _maxHalfAngle = std::numeric_limits<Angle>::infinity();     //<! Maximum off-boresight half angle
+    gtl::btree_map<Angle, Angle> _points; //!< Map of angles defining the polygonal field of view
+    Angle _minHalfAngle = 0.0 * mp_units::angular::unit_symbols::rad; //!< Minimum off-boresight half angle
+    Angle _maxHalfAngle = std::numeric_limits<Angle>::infinity();     //!< Maximum off-boresight half angle
 
     /**
      * @brief Finds the minimum and maximum off-boresight angles from _points
